% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{be_journal_article}
\alias{be_journal_article}
\title{A single journal article as a bibentry object}
\format{
A bibentry object.
}
\source{
\url{https://www.openalex.org}
}
\usage{
be_journal_article
}
\description{
A base R bibentry object with the data for a single journal article.
It is included here for trial and testing purposes.
It could be accessed directly using
}
\keyword{datasets}
