% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibentry.R
\name{bibentry2cp}
\alias{bibentry2cp}
\title{Convert a bibentry (from bibtex) to citeproc csl-data}
\usage{
bibentry2cp(this_bibentry, format = "citeproc-json")
}
\arguments{
\item{this_bibentry}{A bibentry object.}

\item{format}{The format for return of the processed data.}
}
\value{
A object of the selected return \code{format}.
}
\description{
Convert a bibentry (from bibtex) to citeproc csl-data
}
\examples{
\dontrun{
bib <- bibtex::read.bib(package = "base")
bibentry2cp(bib[1])
}
}
