% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cplist2json}
\alias{cplist2json}
\title{Convert citeproc-list data to citeproc-json}
\usage{
cplist2json(l)
}
\arguments{
\item{l}{A list with citeproc formatted data}
}
\value{
A JSON document
}
\description{
Convert citeproc-list data to citeproc-json
}
\examples{
cplist2json(list(title="Hello"))
}
