% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cr_journal_article}
\alias{cr_journal_article}
\title{A single journal article from crossref}
\format{
A list with 43 elements.
}
\source{
\url{https://www.crossref.org}
}
\usage{
cr_journal_article
}
\description{
An R list with the crossref data for a single journal article.
It is included here for trial and testing purposes to avoid repeatedly hitting
crossref servers for no good reason.
It could be accessed directly from crossref using
\preformatted{
rcrossref::cr_cn(dois = "10.1371/journal.pcbi.1009061",
                 format = "citeproc-json")
}
}
\keyword{datasets}
