% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossref.R
\name{crossref2cp}
\alias{crossref2cp}
\title{Convert crossref to citeproc csl-data}
\usage{
crossref2cp(this_data, format = "citeproc-json")
}
\arguments{
\item{this_data}{The data from crossref.}

\item{format}{The format for return of the processed data.}
}
\value{
A object of the selected return \code{format}.
}
\description{
Convert crossref to citeproc csl-data
}
\examples{
\dontrun{
cr_data <- rcrossref::cr_cn(dois = "10.1371/journal.pcbi.1009061",
                            format = "citeproc-json")
crossref2cp(cr_data)
}
}
