% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oa_special_volume_article}
\alias{oa_special_volume_article}
\title{A single journal article in a special volume from openalex}
\format{
A list with 50 elements.
}
\source{
\url{https://www.openalex.org}
}
\usage{
oa_special_volume_article
}
\description{
An R list with the openalex data for a single journal article.
It is included here for trial and testing purposes to avoid repeatedly hitting
openalex servers for no good reason.
It could be accessed directly using
\preformatted{
openalexR::oa_query(entity = "works",
                    doi = "10.1090/conm/355/06452") |>
           openalexR::oa_request()
}
}
\keyword{datasets}
