% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pm_journal_article_xmltext}
\alias{pm_journal_article_xmltext}
\title{A single journal article from pubmed}
\format{
A string of XML data.
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/}
}
\usage{
pm_journal_article_xmltext
}
\description{
XML text with the pubmed data for a single journal article.
It is included here for trial and testing purposes to avoid repeatedly hitting
pubmed servers for no good reason.
It could be accessed directly using
\preformatted{
this_res <- rentrez::entrez_search("pubmed",
                                   term="10.1371/journal.pcbi.1009061")
this_pmid <- this_res$ids[1]
this_record <- rentrez::entrez_fetch("pubmed",
                                     id=this_pmid,
                                     rettype="xml")
}
Note that we cannot save processed XML because the internal R representation
does not permit it, so this example data needs to be parsed before use:
\preformatted{
pm_journal_article <- xml2::read_xml(pm_journal_article_xmltext)
}
}
\keyword{datasets}
