% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed.R
\name{pubmed2cp}
\alias{pubmed2cp}
\title{Convert pubmed xml to citeproc csl-data}
\usage{
pubmed2cp(this_xml, format = "citeproc-json")
}
\arguments{
\item{this_xml}{The xml data from pubmed.}

\item{format}{The format for return of the processed data.}
}
\value{
A object of the selected return \code{format}.
}
\description{
Convert pubmed xml to citeproc csl-data
}
\examples{
\dontrun{
this_res <- entrez_search("pubmed", term="10.1371/journal.pcbi.1009061")
this_pmid <- this_res$ids[1]
this_record <- entrez_fetch("pubmed", id=this_pmid, rettype="xml")
this_xml <- read_xml(this_record)

pubmed2cp(this_xml)
}
}
