% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{xml_text_or_null}
\alias{xml_text_or_null}
\title{Extract text from XML nodeset or return NULL for empty nodeset}
\usage{
xml_text_or_null(x)
}
\arguments{
\item{x}{An xml2 nodeset}
}
\value{
Either text or NULL
}
\description{
Extract text from XML nodeset or return NULL for empty nodeset
}
\examples{
x <- xml2::read_xml("<p>This is some text. This is <b>bold!</b></p>")
xml_text_or_null(x)
}
