% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossref.R
\name{crossref2cp}
\alias{crossref2cp}
\title{Convert crossref to citeproc csl-data}
\usage{
crossref2cp(this_data, format = "list")
}
\arguments{
\item{this_data}{The data from crossref}

\item{format}{The format of the data from crossref, currently accepting only
the list format prepared by the \link{rcrossref} package.}
}
\value{
A list with an \code{item} element and probably an \code{author} (or other creator)
element(s) (i.e. \code{editor}, \code{director}, etc.). Often could have \code{affiliation} and
\code{identifier} elements.
}
\description{
Convert crossref to citeproc csl-data
}
\examples{
\dontrun{
cr_data <- rcrossref::cr_cn(dois = "10.1371/journal.pcbi.1009061",
                            format = "citeproc-json")
crossref2cp(cr_data)
}
}
